Parameters of dispatchRequest should be optional and default to _.noop
The onSuccess, onError and onProgress parameters of dispatchRequest() should be
optional (we might not want to handle these events) and they should default to
_.noop, so that dispatchRequest can call them without triggering a ReferenceError.

This mainly fixes a bug where the default value of onProgress was null and
some handlers didn't specify it. If the request is a POST, the progress event
is fired, leading to a ReferenceError.